+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 
+- Now we use a sidechain LSP compressor that allows the user to select and external source as the sidechain input.
 - When `split-channels` is enabled in the equalizer the imported APO preset will be applied only to the channel being
   visualized in the window. This will allow to import different presets for each channel instead of just settings the same
   values to both.